# runtime/internal/openapidocs/openapi.json
{
  "components": {
    "schemas": {
      "APIError": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "type": {
                "type": "string"
              }
            },
            "type": "object"
          }
        },
        "type": "object"
      },
      "agentinstance_FleetEntry": {
        "properties": {
          "agentId": {
            "type": "string"
          },
          "agentName": {
            "type": "string"
          },
          "instances": {
            "items": {
              "$ref": "#/components/schemas/agentinstance_InstanceStatus"
            },
            "type": "array"
          },
          "kind": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "agentinstance_InstanceStatus": {
        "properties": {
          "agentId": {
            "type": "string"
          },
          "agentName": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "sessionIds": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "sessions": {
            "type": "integer"
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "viewers": {
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "fleetapi_CancelRequest": {
        "properties": {
          "sessionId": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "fleetservice_DispatchRequest": {
        "properties": {
          "agentName": {
            "type": "string"
          },
          "cwd": {
            "type": "string"
          },
          "hitlPolicyName": {
            "type": "string"
          },
          "intent": {
            "type": "string"
          },
          "parentSessionId": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "fleetservice_DispatchResult": {
        "properties": {
          "instanceId": {
            "type": "string"
          },
          "missionId": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP API served by `contenox serve` under /api. Generated from route annotations by tools/openapi-gen.",
    "title": "Contenox Runtime API",
    "version": "v0.36.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/fleet": {
      "get": {
        "operationId": "get_fleet",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/agentinstance_FleetEntry"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "list returns every fleet instance the kernel currently tracks.",
        "tags": [
          "fleet"
        ]
      }
    },
    "/fleet/dispatch": {
      "post": {
        "operationId": "fleet_dispatch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/fleetservice_DispatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/fleetservice_DispatchResult"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "dispatch allocates a unit via fleetservice.Dispatch and returns 202 with the ids as soon as the session is open; the orchestration (Enabled check, teardown-on-failure, the mission record, the detached first turn) lives entirely in the service now.",
        "tags": [
          "fleet"
        ]
      }
    },
...
    "/fleet/{instanceID}": {
      "delete": {
        "operationId": "fleet_stop",
        "parameters": [
          {
            "description": "The unique ID of the instance.",
            "in": "path",
            "name": "instanceID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "stop tears an instance down via fleetservice.Stop, which is idempotent by kernel contract: stopping an unknown or already-stopped instance is a no-op 200, not a 404 — mirrors the delete-response idiom missionapi/mcpserverapi use (a plain \"deleted\" string body).",
        "tags": [
          "fleet"
        ]
      },
      "get": {
        "operationId": "fleet_get",
        "parameters": [
          {
            "description": "The unique ID of the instance.",
            "in": "path",
            "name": "instanceID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/agentinstance_InstanceStatus"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "get returns the live status of one fleet instance.",
        "tags": [
          "fleet"
        ]
      }
    },
    "/fleet/{instanceID}/cancel": {
      "post": {
        "operationId": "fleet_cancel",
        "parameters": [
          {
            "description": "The unique ID of the instance.",
            "in": "path",
            "name": "instanceID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/fleetapi_CancelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "cancel cancels an in-flight prompt turn via fleetservice.Cancel.",
        "tags": [
          "fleet"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}